example modifications to add job options
add queue dropdown, and email text field, and email on job start checkbox

We update ResourceManagerAdapter, our bridge class between the old osc-machete and
the new ood_core library, to accept the workflow it works with as an argument. Thus
ResourceManagerAdapter acts like a service object, and can set the arguments on the
ood_core script object.